+++ CHANGELOG.md
@@ -166,3 +166,7 @@
 #### [(2024-10-18)]
 
 -   Modify filter dropdown in no rekening to make filtering table in index no rekening
+
+#### [(2024-10-21)]
+
+-   Create an event report function, function adjustment anggaran and realisasi